Ne10 FFTs for Neon (#219)
* Neon implementation of the FFTs based upon Ne10 source code
* The API for using FFTs with Neon has changed a little (read documentation)

+++ Testing/Source/Tests/DistanceTestsU32.cpp
@@ -21,7 +21,7 @@
           outp ++;
        }
 
-        ASSERT_REL_ERROR(output,ref,(float32_t)ERROR_THRESHOLD);
+        ASSERT_REL_ERROR(output,ref,ERROR_THRESHOLD);
     } 
 
     void DistanceTestsU32::test_hamming_distance()
